SimpleIME: Draw caret by TextEditor acp selection; try to enable DPI aware in IME thread;

+++ include/configs/AppConfig.h
@@ -152,7 +152,7 @@ public:
         return m_emojiFontFile.Value();
     }
 
-    [[nodiscard]] constexpr auto FontSize() const -> float
+    [[nodiscard]] constexpr auto FontSize() const -> int
     {
         return m_fontSize.Value();
     }
